% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean file list}
\usage{
clean(file_list, ext = ".tif")
}
\arguments{
\item{file_list}{Vector of file paths.}

\item{ext}{File extension to keep.}
}
\description{
Remove thumbnails and other rubbish from filename lists.
}
