% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_img_channel}
\alias{get_img_channel}
\title{return the image channel from the image URL}
\usage{
get_img_channel(img_url, char = "_")
}
\arguments{
\item{img_url}{string, final image URL in ImageXpress file path}

\item{char}{string, separating character in URL}
}
\description{
return the image channel from the image URL
}
