% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_img_filename}
\alias{get_img_filename}
\title{return the final image filename from the path}
\usage{
get_img_filename(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
return the final image filename from the path
}
