% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_img_well}
\alias{get_img_well}
\title{return well from the final image URL}
\usage{
get_img_well(img_url, char = "_")
}
\arguments{
\item{img_url}{string, final image URL in ImageXpress file path}

\item{char}{string, separating character in URL}
}
\description{
return well from the final image URL
}
