% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_path}
\alias{get_path}
\title{return the path up until the image URL}
\usage{
get_path(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
return the path up until the image URL
}
