% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_plate_date}
\alias{get_plate_date}
\title{return the plate date from the image URL}
\usage{
get_plate_date(path, as_datetime = TRUE)
}
\arguments{
\item{path}{string, ImageExpress file path}

\item{as_datetime}{Boolean, whether to convert to a Date object}
}
\description{
return the plate date from the image URL
}
