% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_plate_name}
\alias{get_plate_name}
\title{return the plate name from the image URL}
\usage{
get_plate_name(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
return the plate name from the image URL
}
