% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_plate_num}
\alias{get_plate_num}
\title{return the plate number from the image URL}
\usage{
get_plate_num(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
return the plate number from the image URL
}
