% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{is_full_path}
\alias{is_full_path}
\title{test if path contains file separators}
\usage{
is_full_path(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
test if path contains file separators
}
