% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{split_path}
\alias{split_path}
\title{split path into components by file separator}
\usage{
split_path(path)
}
\arguments{
\item{path}{string, ImageExpress file path}
}
\description{
split path into components by file separator
}
