% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{split_url}
\alias{split_url}
\title{split the image URL into components separated by `char`}
\usage{
split_url(img_url, char = "_")
}
\arguments{
\item{img_url}{string, final image URL in ImageXpress file path}

\item{char}{string, separating character in URL}
}
\description{
split the image URL into components separated by `char`
}
